[feat] #141 - 내 리뷰 불러오기 구현

+++ app/src/main/java/org/helfoome/domain/entity/MyReviewListInfo.kt
@@ -2,7 +2,7 @@ package org.helfoome.domain.entity
 
 data class MyReviewListInfo(
     val id: String,
-    val score: Double,
+    val score: Float,
     val tags: String,
     val good: List<String>,
     val description: String,
